+++ outputs.tf
@@ -116,6 +116,12 @@ output "gcp_us_west1_ips" {
   value       = "${local.gcp-us-west1}"
 }
 
+// Canada/Montreal
+output "gcp_northamerica_northeast1_ips" {
+  description = "Looker IPs for instances hosted in GCP (Canada / Montreal / gcp-northamerica-northeast1)"
+  value       = "${local.gcp-northamerica-northeast1}"
+}
+
 // London
 output "gcp_europe_west2_ips" {
   description = "Looker IPs for instances hosted in GCP (London / europe-west2)"
@@ -135,7 +141,7 @@ output "gcp_asia_southeast1_ips" {
 }
 
 // Sydney
-output "gcp_australia_southeast1" {
+output "gcp_australia_southeast1_ips" {
   description = "Looker IPs for instances hosted in GCP (Sydney / australia-southeast1)"
   value       = "${local.gcp-australia-southeast1}"
 }
@@ -161,6 +167,13 @@ output "gcp_us_west1_cidr" {
   value       = ["${formatlist("%s/32", local.gcp-us-west1)}"]
 }
 
+
+// Canada / Montreal
+output "gcp_northamerica_northeast1_cidr" {
+  description = "Looker CIDRs for instances hosted in GCP (Canada / Montreal / gcp-northamerica-northeast1)"
+  value       = ["${formatlist("%s/32", local.gcp-northamerica-northeast1)}"]
+}
+
 // London
 output "gcp_europe_west2_cidr" {
   description = "Looker CIDRs for instances hosted in GCP (London / europe-west2)"
@@ -180,7 +193,7 @@ output "gcp_asia_southeast1_cidr" {
 }
 
 // Singapore
-output "gcp_australia_southeast1" {
+output "gcp_australia_southeast1_cidr" {
   description = "Looker CIDRs for instances hosted in GCP (Sydney / australia-southeast1)"
   value       = ["${formatlist("%s/32", local.gcp-australia-southeast1)}"]
 }
@@ -194,6 +207,7 @@ output "gcp_cidr" {
     local.gcp-us-east1,
     local.gcp-us-east4,
     local.gcp-us-west1,
+    local.gcp-northamerica-northeast1,
     local.gcp-europe-west2,
     local.gcp-europe-west3,
     local.gcp-asia-southeast1,
@@ -209,6 +223,7 @@ output "gcp_ips" {
     local.gcp-us-east1,
     local.gcp-us-east4,
     local.gcp-us-west1,
+    local.gcp-northamerica-northeast1,
     local.gcp-europe-west2,
     local.gcp-europe-west3,
     local.gcp-asia-southeast1,
@@ -225,6 +240,7 @@ output "all_cidr" {
     local.gcp-us-east1,
     local.gcp-us-east4,
     local.gcp-us-west1,
+    local.gcp-northamerica-northeast1,
     local.gcp-europe-west2,
     local.gcp-europe-west3,
     local.gcp-asia-southeast1,
@@ -247,6 +263,7 @@ output "all_ips" {
     local.gcp-us-east1,
     local.gcp-us-east4,
     local.gcp-us-west1,
+    local.gcp-northamerica-northeast1,
     local.gcp-europe-west2,
     local.gcp-europe-west3,
     local.gcp-asia-southeast1,