+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.3][1.2.3] - 2020-06-21
+
+- New Looker IP for GCP/Canada region
+- Fixes output variable name for Sydney regions
+
 ## [1.2.2][1.2.2] - 2020-06-13
 
 ### Added